requirements.txt
flask
instaloader
opencv-python-headless
google-generativeai
python-dotenv